voice_ai/requirements.txt
spotipy
SpeechRecognition
python-dotenv
requests
